+++ src/services/traits/records/ActiveRecord.php
@@ -36,6 +36,36 @@ trait ActiveRecord
     }
 
     /*******************************************
+     * CREATE
+     *******************************************/
+
+    /**
+     * @param array $attributes
+     * @param string $toScenario
+     * @return Record
+     */
+    public function create(array $attributes = [], string $toScenario = null): Record
+    {
+        /** @var string $recordClass */
+        $recordClass = static::recordClass();
+
+        /** @var Record $record */
+        $record = new $recordClass();
+
+        // Set scenario
+        if ($toScenario) {
+            $record->setScenario($toScenario);
+        }
+
+        // Do we need to set properties too
+        if (!empty($attributes)) {
+            $record->setAttributes($attributes);
+        }
+
+        return $record;
+    }
+
+    /*******************************************
      * QUERY
      *******************************************/