+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+## Unreleased
+### Added
+- ActiveRecord::create trait method to create a new ActiveRecord
+
 ## 1.0.0-rc.14 - 2018-03-01
 ### Changed
 - Legacy translation categories